reconfigures handling of pages to remove routing
