+++ host_v2/src/App.tsx
@@ -16,26 +16,14 @@ function RedirectToPlayIfMissing() {
   window.location.href = 'http://dev-central.rightoneducation.com/';
   return null;
 }
-const mockGameSession = GameSessionParser.gameSessionFromAWSGameSession({
-  ...MockGameSession,
-  currentState: MockGameSession.currentState as GameSessionState,
-});
-const handleStartGame = ()=>{
-  console.log("test")
-}
+
 const apiClient = new ApiClient(Environment.Developing);
 const router = createBrowserRouter(
   createRoutesFromElements(
     <>
       <Route
         path="/"
-        
         element={<GameSessionContainer apiClient={apiClient} />}
-      />
-      <Route
-        path="/StartGame"
-        element={<StartGame teams={mockGameSession.teams ?? []} currentQuestionIndex={mockGameSession.currentQuestionIndex ?? 0} questions={mockGameSession.questions} title={mockGameSession.title ?? ''} gameSessionId={mockGameSession.id} 
-        gameCode={mockGameSession.gameCode ?? 1100} currentState={mockGameSession.currentState} handleStartGame={handleStartGame} />}
       />
       <Route element={<RedirectToPlayIfMissing />} />
     </>,
